api/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
openai==1.18.0
pydantic>=2.0.0
typing-extensions>=4.7.0
httpx==0.27.2
requests==2.31.0
pyjwt==2.8.0
cryptography==42.0.0
redis==5.0.1 